# Minor performance fixes