hello @bbakerman @andimarek :)

another batch of small fixes... please have a look when you have time :)

D.